Refined Group Service

+++ src/main/java/com/contacts/Entity/Users.java
@@ -26,11 +26,11 @@ import lombok.Data;
 @Data
 public class Users {
 
-    public int getId() {
+    public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -100,7 +100,7 @@ public class Users {
 
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(nullable = false)
     private String firstName;
